MANY Changes
mainly about the test-train split, i've so far fixed sca preprocessor and the baseline scripts.

but the upload here is actually for the kmcluster elbow plot. So I can look at it.

I've written a batch script, even though its just the kmcluster script called with the --elbow argument. let's see if it works